Move extension scripts to own folder
