+++ src/scripts/chat-background.ts
@@ -1,7 +1,7 @@
-import type { Unsubscriber } from './queue';
-import { isValidFrameInfo } from './chat-utils';
-import { isLiveTL } from './chat-constants';
-import { ytcQueue } from './queue';
+import type { Unsubscriber } from '../ts/queue';
+import { isValidFrameInfo } from '../ts/chat-utils';
+import { isLiveTL } from '../ts/chat-constants';
+import { ytcQueue } from '../ts/queue';
 
 const interceptors: Chat.Interceptors[] = [];
